refactor(api): déplacer les services de commentaires de site vers shantytownComment/ refers to https://trello.com/c/evsOelCU/2717
deleteComment.ts et enrichCommentsAttachments.ts vivaient encore dans
services/shantytown/, aux côtés du CRUD des sites, alors que createComment
et updateComment étaient déjà isolés dans services/shantytownComment/.
Alignement sur le même pattern que le domaine action (services/actionComment/) :
mise à jour des barrels shantytown/index.ts et shantytownComment/index.ts,
et du seul consommateur (shantytown.moderateComment.ts) pour importer
deleteComment depuis le nouveau service.
