+++ packages/api/server/services/shantytown/index.ts
@@ -7,7 +7,6 @@ import forceUpdateWithoutChanges from './forceUpdateWithoutChanges';
 import getClosureYearRange from './getClosureYearRange';
 import close from './close';
 import deleteTown from './deleteTown';
-import deleteComment from './deleteComment';
 import exportTowns from './exportTowns';
 import fixClosedStatus from './fixClosedStatus';
 import setHeatwaveStatus from './setHeatwaveStatus';
@@ -24,7 +23,6 @@ export default {
     getClosureYearRange,
     close,
     deleteTown,
-    deleteComment,
     exportTowns,
     fixClosedStatus,
     setHeatwaveStatus,
