+++ packages/api/server/services/shantytownComment/index.ts
@@ -1,9 +1,11 @@
 import createComment from './createComment';
+import deleteComment from './deleteComment';
 import exportAll from './exportAll';
 import updateComment from './updateComment';
 
 export default {
     createComment,
+    deleteComment,
     exportAll,
     updateComment,
 };
